(feat-dev): Mostrando código da sala
