+++ package.json
@@ -16,6 +16,7 @@
     "@mui/material": "^9.0.0",
     "@tailwindcss/vite": "^4.2.2",
     "axios": "^1.15.0",
+    "qrcode": "^1.5.4",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
     "react-router-dom": "^7.14.1",
