+++ src/pages/room/[id]/page.tsx
@@ -1,11 +1,7 @@
 import { Navbar } from "../../../components/Navbar";
 import { useRoom } from "../../../hooks/useRoom";
 import { ButtonGroup } from "../../../components/ButtonGroup/index";
-import { useState, useEffect } from "react";
-import { RoomForm } from '../../../components/RoomForm/index';
-import { Input } from "../../../components/Input/index";
-import { Tooltip } from "@mui/material";
-import { strings, requiredFieldString, queueString, participantsString, dataString } from "../../../utils/strings";
+import { strings, queueString, participantsString, dataString, roomCodeString } from "../../../utils/strings";
 import { language } from "../../../utils/settings";
 import { RoomData } from "./RoomData";
 
@@ -13,28 +9,30 @@ export function RoomPage() {
 
   const { 
     room,
-    name,
-    setName,
-    nameError,
-    setNameError,
-    password,
-    setPassword,
-    maxQuantity,
-    setMaxQuantity,
-    maxQuantityError,
-    setMaxQuantityError,
     activeButton, 
     setActiveButton,
-    handleClose,
     handleEdit,
-    validations,
+    qrCodeUrl,
   } = useRoom();
 
   return (
     room ? 
     <div className="flex flex-col p-14 min-h-screen"> 
       <Navbar/>
-      <div className="flex items-center justify-center text-[1cm] h-20">{room.name}</div> 
+      <div className="flex items-center justify-between text-[1cm] mx-20 h-20">
+        <div>
+          {room.name}
+        </div>
+        <div>
+          {`${strings[language][roomCodeString]}: ${room.code}`}
+          {qrCodeUrl && (
+            <img
+              src={qrCodeUrl}
+              className="inline-block ml-4 w-24 h-15 shadow-sm border border-gray-200"
+            />
+          )}
+        </div>
+      </div> 
       <ButtonGroup 
         buttonsText={[strings[language][queueString], strings[language][participantsString], strings[language][dataString]]}
         activeButtonText={activeButton}
@@ -46,19 +44,7 @@ export function RoomPage() {
         {activeButton === strings[language][dataString] && 
           <RoomData
             room={room}
-            name={name}
-            setName={setName}
-            nameError={nameError}
-            setNameError={setNameError}
-            password={password}
-            setPassword={setPassword}
-            maxQuantity={maxQuantity}
-            setMaxQuantity={setMaxQuantity}
-            maxQuantityError={maxQuantityError}
-            setMaxQuantityError={setMaxQuantityError}
-            handleClose={handleClose}
             handleEdit={handleEdit}
-            validations={validations}
           />
         }
       </div>
