+++ src/utils/settings.ts
@@ -1,2 +1,3 @@
 export const defaultLanguage: string = "pt-br"
 export const language: string = defaultLanguage;
+export const url = import.meta.env.VITE_FRONT_URL;
